Get file types

Get file types based on the file's content.

file_type(files, mime_type = FALSE, mime_encoding = FALSE,
  magicfiles = NULL, names = FALSE)

Arguments

files
Vector of file names including full or relative paths

mime_type
Logical indicating whether the MIME type should be returned

mime_encoding
Logical indicating whether the MIME encoding should be returned

magicfiles
Vector with file or directory names to be used as magic file; the first usable location is selected

names
Logical indicating whether the file names should be added to the result as attribute names

Value

Vector of textual file type descriptions or MIME types/encodings